[hotfix] Fixed an issue where overflow from Long to Int was unsafe

+++ core/src/main/java/org/apache/calcite/rex/RexInterpreter.java
@@ -213,7 +213,7 @@ public class RexInterpreter implements RexVisitor<Comparable> {
     final int v2;
     if (v instanceof Long) {
       // TIMESTAMP
-      v2 = (int) (((Long) v) / TimeUnit.DAY.multiplier.longValue());
+      v2 = Math.toIntExact(((Long) v) / TimeUnit.DAY.multiplier.longValue());
     } else {
       // DATE
       v2 = (Integer) v;
